Seperated code for firehose logging
